Run platform sanity checks after creating the log and log windows path
